+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Tidarr notable changes.
 - [API] Tidal playlist fetches now renew token if needed
 - [Lidarr] Apply PUID and PGID to Lidarr downloaded files
 - [Front] Add single download button in no-download mode #684
+- [Front] Add toggle finished items button in queue list
 
 ## 📦 1.2.0
 ### 🐛 Fixed